Move custom commands into their own Commands settings section
